Files given through the command line are now accepted properly. Memory view now remembers where it left off before.

+++ includes/Memory.h
@@ -3,11 +3,10 @@
 
 #include <curses.h>
 
-#include "Enums.h"
 #include "Structs.h"
+#include "Enums.h"
 
-extern int selected;
-
+extern uint16_t selected;
 extern uint16_t output_height;
 extern uint16_t *memory_output;
 extern uint16_t selected_address;
